refactor(kilo-chat): share attachment helpers
