+++ packages/kilo-chat-hooks/src/use-attachment-queue.ts
@@ -14,6 +14,13 @@ export type QueuedAttachment = {
   error?: string;
 };
 
+type XhrUploadResult = { status: number; aborted: boolean };
+
+export type XhrUploadOutcome =
+  | { kind: 'ok' }
+  | { kind: 'aborted' }
+  | { kind: 'error'; message: string };
+
 export type AttachmentQueueState = { rows: QueuedAttachment[] };
 
 export type AttachmentQueueAction =
@@ -127,6 +134,84 @@ export type PerformUpload = (
   putHeaders: Record<string, string>,
   opts: { onProgress: (fraction: number) => void; signal: AbortSignal }
 ) => Promise<void>;
+
+export function mapXhrUploadResultToOutcome(result: XhrUploadResult): XhrUploadOutcome {
+  if (result.aborted) {
+    return { kind: 'aborted' };
+  }
+
+  if (result.status === 0) {
+    return { kind: 'error', message: 'Network error during upload' };
+  }
+
+  if (result.status >= 200 && result.status < 300) {
+    return { kind: 'ok' };
+  }
+
+  return { kind: 'error', message: `Upload failed (${result.status})` };
+}
+
+export function createXhrPerformUpload(): PerformUpload {
+  return (blob, putUrl, putHeaders, opts) =>
+    new Promise<void>((resolve, reject) => {
+      const xhr = new XMLHttpRequest();
+      let aborted = false;
+
+      function cleanup() {
+        opts.signal.removeEventListener('abort', onAbort);
+      }
+
+      function createAbortError(): DOMException {
+        return new DOMException('Aborted', 'AbortError');
+      }
+
+      function onAbort() {
+        aborted = true;
+        try {
+          xhr.abort();
+        } catch {
+          // Ignore abort errors from native XHR cleanup.
+        }
+        cleanup();
+        reject(createAbortError());
+      }
+
+      if (opts.signal.aborted) {
+        onAbort();
+        return;
+      }
+
+      opts.signal.addEventListener('abort', onAbort, { once: true });
+      xhr.open('PUT', putUrl, true);
+
+      for (const [key, value] of Object.entries(putHeaders)) {
+        if (key.toLowerCase() !== 'content-length') {
+          xhr.setRequestHeader(key, value);
+        }
+      }
+
+      xhr.upload.addEventListener('progress', event => {
+        if (!event.lengthComputable || event.total === 0) {
+          return;
+        }
+        opts.onProgress(event.loaded / event.total);
+      });
+
+      xhr.addEventListener('loadend', () => {
+        cleanup();
+        const outcome = mapXhrUploadResultToOutcome({ status: xhr.status, aborted });
+        if (outcome.kind === 'ok') {
+          resolve();
+        } else if (outcome.kind === 'aborted') {
+          reject(createAbortError());
+        } else {
+          reject(new Error(outcome.message));
+        }
+      });
+
+      xhr.send(blob);
+    });
+}
 
 export type AddFileInput = {
   blob: Blob;
