+++ packages/kilo-chat/src/index.ts
@@ -5,6 +5,9 @@ export {
   contentBlocksToText,
   contentBlocksPreviewText,
   buildReplyToMessageSnapshot,
+  formatFileSize,
+  remainingEditableAttachments,
+  buildMessageEditContent,
   encodeConversationCursor,
   decodeConversationCursor,
   type ConversationCursor,
