+++ packages/kilo-chat/src/utils.ts
@@ -2,11 +2,58 @@ import { decodeTime } from 'ulid';
 import type { z } from 'zod';
 
 import { conversationCursorSchema } from './schemas';
-import type { ReplyToMessageSnapshot } from './types';
+import type { AttachmentBlock, InputContentBlock, ReplyToMessageSnapshot } from './types';
+
+const FILE_SIZE_UNITS = ['B', 'KB', 'MB', 'GB'] as const;
 
 /** Extract the millisecond timestamp encoded in a ULID. */
 export function ulidToTimestamp(ulid: string): number {
   return decodeTime(ulid);
+}
+
+export function formatFileSize(bytes: number): string {
+  let value = bytes;
+  let unitIndex = 0;
+
+  while (value >= 1024 && unitIndex < FILE_SIZE_UNITS.length - 1) {
+    value /= 1024;
+    unitIndex += 1;
+  }
+
+  if (unitIndex === 0) {
+    return `${Math.round(value)} ${FILE_SIZE_UNITS[unitIndex]}`;
+  }
+
+  const rounded = Math.round(value * 100) / 100;
+  return `${rounded} ${FILE_SIZE_UNITS[unitIndex]}`;
+}
+
+export function remainingEditableAttachments(
+  originalAttachments: readonly AttachmentBlock[],
+  removedAttachmentIds: Iterable<string>
+): AttachmentBlock[] {
+  const removedAttachmentIdSet = new Set(removedAttachmentIds);
+  return originalAttachments.filter(
+    attachment => !removedAttachmentIdSet.has(attachment.attachmentId)
+  );
+}
+
+export function buildMessageEditContent({
+  text,
+  originalAttachments,
+  removedAttachmentIds,
+}: {
+  text: string;
+  originalAttachments: readonly AttachmentBlock[];
+  removedAttachmentIds: Iterable<string>;
+}): InputContentBlock[] {
+  const trimmedText = text.trim();
+  const textBlocks: InputContentBlock[] =
+    trimmedText.length > 0 ? [{ type: 'text', text: trimmedText }] : [];
+  return [
+    ...textBlocks,
+    ...remainingEditableAttachments(originalAttachments, removedAttachmentIds),
+  ];
 }
 
 /**
