+++ packages/kilo-chat/test/format-file-size.test.ts
@@ -1,20 +1,22 @@
-import { formatFileSize } from './format-file-size';
+import { describe, expect, it } from 'vitest';
+
+import { formatFileSize } from '../src';
 
 describe('formatFileSize', () => {
-  it('formats bytes', () => {
+  it('formats bytes without decimal places', () => {
     expect(formatFileSize(0)).toBe('0 B');
     expect(formatFileSize(512)).toBe('512 B');
     expect(formatFileSize(1023)).toBe('1023 B');
   });
-  it('formats KB', () => {
+
+  it('formats kilobytes with up to two decimal places', () => {
     expect(formatFileSize(1024)).toBe('1 KB');
     expect(formatFileSize(1536)).toBe('1.5 KB');
   });
-  it('formats MB', () => {
+
+  it('formats megabytes with up to two decimal places', () => {
     expect(formatFileSize(1024 * 1024)).toBe('1 MB');
     expect(formatFileSize(5.25 * 1024 * 1024)).toBe('5.25 MB');
-  });
-  it('rounds to at most 2 decimals', () => {
     expect(formatFileSize(1024 * 1024 + 7777)).toBe('1.01 MB');
   });
 });
